Moved encodeWebAuthnSigningMessage to test
