+++ modules/passkey/test/SafeWebAuthnSigner.spec.ts
@@ -2,7 +2,8 @@ import { expect } from 'chai'
 import { deployments, ethers } from 'hardhat'
 
 import * as ERC1271 from './utils/erc1271'
-import { DUMMY_AUTHENTICATOR_DATA, base64UrlEncode, encodeWebAuthnSigningMessage, getSignatureBytes } from '../src/utils/webauthn'
+import { DUMMY_AUTHENTICATOR_DATA, base64UrlEncode, getSignatureBytes } from '../src/utils/webauthn'
+import { encodeWebAuthnSigningMessage } from './utils/webauthn'
 
 describe('SafeWebAuthnSigner', () => {
   const setupTests = deployments.createFixture(async () => {
